search and search filters

+++ app/src/main/java/com/spc/space/models/bookingsHistory/BookingHistoryResponse.kt
@@ -11,10 +11,10 @@ data class History(
 //    val createdAt: String,
 //    val dateCreated: String,
     @SerializedName("_id") val id: String,
-    val duration: Int,
+    val duration: Float,
     val endTime: String,
     val isCancelled: Boolean,
-    val price: Int,
+    val price: Float,
     val room: Room,
     val startTime: String,
 )
